+++ src/app/vehicule/vehicule-list/vehicule-list.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit, Output} from '@angular/core';
 import {Observable} from 'rxjs';
 import {Router} from '@angular/router';
 import {VehiculeService} from '../vehicule.service';
@@ -24,13 +24,15 @@ export class VehiculeListComponent implements OnInit {
   }
 
   deleteVehicule(id: number): void {
-    this.vehiculeService.deleteVehicule(id)
-      .subscribe(
-        data => {
-          console.log(data);
-          this.reloadData();
-        },
-        error => console.log(error));
+    if(confirm('Supprimer le véhicule ?')) {
+      this.vehiculeService.deleteVehicule(id)
+        .subscribe(
+          data => {
+            console.log(data);
+            this.reloadData();
+          },
+          error => console.log(error));
+    }
   }
 
   vehiculeDetails(id: number): void {
